use signed int to fit into Postgres integer type

+++ file/file.go
@@ -50,7 +50,7 @@ type File struct {
 	Rank int
 
 	// CRC32 checksum of the file content
-	Checksum uint32
+	Checksum int32
 }
 
 // Files is a slice of Files
@@ -79,7 +79,8 @@ func (f *File) ReadContent() error {
 			return err
 		}
 		f.Content = content
-		f.Checksum = crc32.ChecksumIEEE(content)
+		// Convert to signed int for Postgres integer type
+		f.Checksum = int32(crc32.ChecksumIEEE(content))
 	}
 	return nil
 }
